+++ src/progress.rs
@@ -72,15 +72,20 @@ pub struct WorkDoneProgressBegin {
 
     /// Optional, more detailed associated progress message. Contains
     /// complementary information to the `title`.
+    ///
     /// Examples: "3/25 files", "project/src/module2", "node_modules/some_dep".
     /// If unset, the previous progress message (if any) is still valid.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub message: Option<String>,
 
     /// Optional progress percentage to display (value 100 is considered 100%).
-    /// If unset, the previous progress percentage (if any) is still valid.
+    /// If not provided infinite progress is assumed and clients are allowed
+    /// to ignore the `percentage` value in subsequent in report notifications.
+    ///
+    /// The value should be steadily rising. Clients are free to ignore values
+    /// that are not following this rule. The value range is [0, 100]
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub percentage: Option<f64>,
+    pub percentage: Option<u32>,
 }
 
 #[derive(Debug, PartialEq, Default, Deserialize, Serialize, Clone)]
@@ -100,9 +105,13 @@ pub struct WorkDoneProgressReport {
     pub message: Option<String>,
 
     /// Optional progress percentage to display (value 100 is considered 100%).
-    /// If unset, the previous progress percentage (if any) is still valid.
+    /// If not provided infinite progress is assumed and clients are allowed
+    /// to ignore the `percentage` value in subsequent in report notifications.
+    ///
+    /// The value should be steadily rising. Clients are free to ignore values
+    /// that are not following this rule. The value range is [0, 100]
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub percentage: Option<f64>,
+    pub percentage: Option<u32>,
 }
 
 #[derive(Debug, PartialEq, Default, Deserialize, Serialize, Clone)]
